+++ src/Definition/Type/PointInTimeType.php
@@ -55,12 +55,12 @@ final class PointInTimeType implements Type
             ->map(static fn() => new self($clock, new Format));
     }
 
-    public function normalize(mixed $value): null|string|int|bool
+    public function normalize(mixed $value): null|string|int|float|bool
     {
         return $value->format($this->format);
     }
 
-    public function denormalize(null|string|int|bool $value): mixed
+    public function denormalize(null|string|int|float|bool $value): mixed
     {
         if (!\is_string($value)) {
             throw new \LogicException("'$value' is not a string");