+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 4.0.0 - 2024-10-31
+
+### Added
+
+- Allow to convert types to floating points.
+
+### Fixed
+
+- Cross matching on an aggregate or entity property.
+
 ## 3.4.1 - 2024-10-26
 
 ### Changed